requirements.txt
anyio==4.3.0
astor==0.8.1
attrdict==2.0.1
attrs==23.2.0
Automat==22.10.0
Babel==2.15.0
bce-python-sdk==0.9.9
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
constantly==23.10.4
contourpy==1.2.1
cryptography==42.0.7
cssselect==1.2.0
cssutils==2.11.0
cycler==0.12.1
Cython==3.0.10
decorator==5.1.1
defusedxml==0.7.1
et-xmlfile==1.1.0
filelock==3.14.0
fire==0.6.0
Flask==3.0.3
flask-babel==4.0.0
Flask-Cors==4.0.1
Flask-Reuploaded==1.4.0
fonttools==4.51.0
future==1.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
hyperlink==21.0.0
idna==3.7
imageio==2.34.1
imgaug==0.4.0
incremental==22.10.0
itemadapter==0.9.0
itemloaders==1.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
kiwisolver==1.4.5
lazy_loader==0.4
lmdb==1.4.1
lxml==5.2.2
MarkupSafe==2.1.5
matplotlib==3.8.4
networkx==3.3
numpy==1.26.4
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.9.0.80
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==24.0
paddleocr==2.7.3
paddlepaddle==2.6.1
pandas==2.2.2
parsel==1.9.1
pdf2docx==0.5.8
pillow==10.3.0
premailer==3.10.0
Protego==0.3.1
protobuf==3.20.2
psutil==5.9.8
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyclipper==1.3.0.post5
pycparser==2.22
pycryptodome==3.20.0
PyDispatcher==2.0.7
PyMuPDF==1.24.3
PyMuPDFb==1.24.3
pyOpenSSL==24.1.0
pyparsing==3.1.2
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.1
PyYAML==6.0.1
queuelib==1.7.0
rapidfuzz==3.9.0
rarfile==4.2
requests==2.31.0
requests-file==2.0.0
scikit-image==0.23.2
scipy==1.13.0
Scrapy==2.11.2
service-identity==24.1.0
setuptools==69.5.1
shapely==2.0.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
termcolor==2.4.0
tifffile==2024.5.10
tldextract==5.1.2
tqdm==4.66.4
Twisted==24.3.0
twisted-iocpsupport==1.0.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
visualdl==2.5.3
w3lib==2.1.2
Werkzeug==3.0.3
wheel==0.43.0
zope.interface==6.4